feat: improve firebase increment points

+++ app/src/main/java/com/neo/hash/data/response/Points.kt
@@ -4,9 +4,9 @@ import com.google.firebase.database.PropertyName
 
 class Points {
     @PropertyName("hard")
-    var hard : Int = 0
+    var hard : Long = 0
     @PropertyName("medium")
-    var medium : Int = 0
+    var medium : Long = 0
     @PropertyName("easy")
-    var easy : Int = 0
+    var easy : Long = 0
 }
